+++ src/Rocket3d/input/AppMessages.cpp
@@ -1,5 +1,5 @@
 #include "AppMessages.h"
-#include "../../../core/Application.h"
+#include "../core/Application.h"
 #include <logging/RK_Log.h>
 
 AppKeyDown::AppKeyDown(const int& key)
